requirements.txt
# Nicholas J Uhlhorn
# November 2025

flask
flask_sqlalchemy 
